+++ soda/cmd/generate/config.go
@@ -2,11 +2,9 @@ package generate
 
 import (
 	"os"
-	"path"
 	"path/filepath"
 	"strings"
 
-	"github.com/gobuffalo/envy"
 	"github.com/gobuffalo/makr"
 	"github.com/markbates/going/defaults"
 	"github.com/pkg/errors"
@@ -25,43 +23,25 @@ var ConfigCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		cflag := cmd.Flag("config")
 		cfgFile := defaults.String(cflag.Value.String(), "database.yml")
-		dir, err := os.Getwd()
+		pwd, err := os.Getwd()
 		if err != nil {
 			return errors.Wrap(err, "couldn't get the current directory")
 		}
-		pwd, _ := os.Getwd()
 		data := map[string]interface{}{
 			"dialect": dialect,
-			"name":    path.Base(dir),
-			"appPath": pwd,
+			"name":    filepath.Base(pwd),
 		}
 		return GenerateConfig(cfgFile, data)
 	},
 }
 
-func goPath(root string) string {
-	gpMultiple := envy.GoPaths()
-	path := ""
-
-	for i := 0; i < len(gpMultiple); i++ {
-		if strings.HasPrefix(root, filepath.Join(gpMultiple[i], "src")) {
-			path = gpMultiple[i]
-			break
-		}
-	}
-	return path
-}
-
-func packagePath(rootPath string) string {
-	gosrcpath := strings.Replace(filepath.Join(goPath(rootPath), "src"), "\\", "/", -1)
-	rootPath = strings.Replace(rootPath, "\\", "/", -1)
-	return strings.Replace(rootPath, gosrcpath+"/", "", 2)
-}
-
 func GenerateConfig(cfgFile string, data map[string]interface{}) error {
+	pwd, _ := os.Getwd()
 	if data["appPath"] == nil {
-		pwd, _ := os.Getwd()
 		data["appPath"] = pwd
+	}
+	if data["sqlitePath"] == nil {
+		data["sqlitePath"] = pwd
 	}
 
 	dialect = strings.ToLower(data["dialect"].(string))